ai_backend/requirements.txt
fastapi
uvicorn
pymongo
langgraph
faiss-cpu
tavily-python
google-generativeai
pydantic
python-dotenv
langchain
langchain_google_genai